# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Work Stat Lite Api Documentation",
        "description": "Work Stat Lite Api Documentation",
        "contact": {
            "email": "orutu1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://wslite-backend.herokuapp.com/api//",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Get list of all employees",
                "description": "Returns list of employees",
                "operationId": "GetEmployees",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/employees-json": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Onboard employees",
                "description": "Onboard employees",
                "operationId": "Onboard employees",
                "parameters": [
                    {
                        "name": "employees",
                        "in": "query",
                        "description": "Employee json array",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/employees-register": {
            "post": {
                "tags": [
                    "Employee"
                ],
                "summary": "Register",
                "description": "Register employee",
                "operationId": "postEmployee",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "query",
                        "description": "Employee id",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Employee name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Employee email address"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Employee phone number"
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Employee gender"
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "description": "Employee department"
                    },
                    {
                        "name": "designation",
                        "in": "query",
                        "description": "Employee designation"
                    },
                    {
                        "name": "image_path",
                        "in": "query",
                        "description": "Employee image_path"
                    },
                    {
                        "name": "appointment_date",
                        "in": "query",
                        "description": "Appointment date"
                    },
                    {
                        "name": "termination_date",
                        "in": "query",
                        "description": "Termination date"
                    },
                    {
                        "name": "card_no",
                        "in": "query",
                        "description": "Card number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Employee created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/create-image": {
            "post": {
                "tags": [
                    "Employee image path created"
                ],
                "summary": "create image paht",
                "description": "create iamge path for employees",
                "operationId": "crete image path",
                "parameters": [
                    {
                        "name": "image_path",
                        "in": "query",
                        "description": "image path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/employee-show": {
            "get": {
                "tags": [
                    "Employee"
                ],
                "summary": "Get employee details",
                "description": "Returns employee details",
                "operationId": "getEmployeeDetails",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/employee-update": {
            "patch": {
                "tags": [
                    "Employee"
                ],
                "summary": "Update employee records",
                "description": "Update employee records",
                "operationId": "updateEmployee",
                "parameters": [
                    {
                        "name": "employee_id",
                        "in": "query",
                        "description": "Employee id",
                        "required": true
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Employee name",
                        "required": true
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Employee email address"
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Employee phone number"
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Employee gender"
                    },
                    {
                        "name": "department",
                        "in": "query",
                        "description": "Employee department"
                    },
                    {
                        "name": "designation",
                        "in": "query",
                        "description": "Employee designation"
                    },
                    {
                        "name": "image_path",
                        "in": "query",
                        "description": "Employee image_path"
                    },
                    {
                        "name": "appointment_date",
                        "in": "query",
                        "description": "Appointment date"
                    },
                    {
                        "name": "termination_date",
                        "in": "query",
                        "description": "Termination date"
                    },
                    {
                        "name": "card_no",
                        "in": "query",
                        "description": "Card number"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/profile-image": {
            "patch": {
                "tags": [
                    "Employee"
                ],
                "summary": "Add profile picture",
                "description": "Add profile picture",
                "operationId": "Add profile picture",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Employee id",
                        "required": true
                    },
                    {
                        "name": "image_path",
                        "in": "query",
                        "description": "Employee id",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/employee-delete": {
            "delete": {
                "tags": [
                    "Employee"
                ],
                "summary": "Delete employee records",
                "description": "Delete employee records",
                "operationId": "deleteEmployee",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}